+++ src/parser/databases/address_index_to_empty_address_data.rs
@@ -7,12 +7,11 @@ use std::{
 
 use allocative::Allocative;
 use itertools::Itertools;
+use snkrj::{AnyDatabase, Database as _Database};
 
 use crate::structs::{Config, EmptyAddressData};
 
-use super::{
-    AnyDatabase, AnyDatabaseGroup, Database as _Database, Metadata, ADDRESS_INDEX_DB_MAX_SIZE,
-};
+use super::{AnyDatabaseGroup, Metadata, ADDRESS_INDEX_DB_MAX_SIZE};
 
 type Key = u32;
 type Value = EmptyAddressData;
@@ -22,6 +21,7 @@ type Database = _Database<Key, Value>;
 pub struct AddressIndexToEmptyAddressData {
     path: PathBuf,
     pub metadata: Metadata,
+    #[allocative(skip)]
     map: BTreeMap<usize, Database>,
 }
 
@@ -40,10 +40,10 @@ impl DerefMut for AddressIndexToEmptyAddressData {
 }
 
 impl AddressIndexToEmptyAddressData {
-    pub fn unsafe_insert(&mut self, key: Key, value: Value) -> Option<Value> {
+    pub fn insert_to_ram(&mut self, key: Key, value: Value) -> Option<Value> {
         self.metadata.called_insert();
 
-        self.open_db(&key).unsafe_insert(key, value)
+        self.open_db(&key).insert_to_ram(key, value)
     }
 
     pub fn remove(&mut self, key: &Key) -> Option<Value> {
@@ -54,13 +54,13 @@ impl AddressIndexToEmptyAddressData {
 
     /// Doesn't check if the database is open contrary to `safe_get` which does and opens if needed
     /// Though it makes it easy to use with rayon.
-    pub fn unsafe_get_from_cache(&self, key: &Key) -> Option<&Value> {
+    pub fn get_from_ram(&self, key: &Key) -> Option<&Value> {
         let db_index = Self::db_index(key);
 
-        self.get(&db_index).and_then(|db| db.get_from_puts(key))
+        self.get(&db_index).and_then(|db| db.get_from_ram(key))
     }
 
-    pub fn unsafe_get_from_db(&self, key: &Key) -> Option<&Value> {
+    pub fn get_from_disk(&self, key: &Key) -> Option<&Value> {
         let db_index = Self::db_index(key);
 
         self.get(&db_index)
@@ -68,7 +68,7 @@ impl AddressIndexToEmptyAddressData {
                 dbg!(&self.map.keys(), &key, &db_index);
                 panic!()
             })
-            .db_get(key)
+            .get_from_disk(key)
     }
 
     pub fn open_db(&mut self, key: &Key) -> &mut Database {